+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+[2022-3-27] 1.2
+- Added automatic ngrok downloader
+- Fixed type errors
+- Changed util to bin
+- Removed ngrok module
+
 [2022-3-27] 1.1
 - Added Linux and Termux support
 - Added automatic module downloader